Actually make deployment work. (Now it works like a charm. I tested it myself!)
The bug was that extra files need to be referenced in the project as
Content items. Along the way, I discovered that HandbrakeCLI.exe is
free-standing, and so I don't need all the rest of that crap. YAY!

Closes #4. For reals, this time. I really need to test harder before I
mark these things closed. If I were feeling particularly vain, I'd go
into GIT and clean up the commit history so that it looks like I got
this right the first time. But I'm too lazy for that.