Use the autocomplete box from the WPF toolkit.
We're taking the autocomplete box because it has better behavior
overall than mine. However, this one needs a change in the system: we
need to put the release year in with the title, because otherwise we
can't distinguish between identical releases.

I think this is better behaviour, so we're going to start moving
towards it.