Fix it so that you can drag the window by clicking anywhere that isn't a control. (Fixes #26)
We accomplish this by moving away from MetroWindow and handing
movement ourselves; in addition, we re-work the TabControl template so
that the top row of tabs does not eat up the entire space; this allows
us to adequately capture MouseDown events in that strip. That strip is
important to us because it lives at the top of the window, where our
title bar would otherwise go.